Updated Header onClick handlers to modify context vars directly instead of through prop functions
